python-bot/requirements.txt
python-telegram-bot>=20.0
python-telegram-bot[job-queue]
requests
python-dotenv
selenium